dual_choice/pyproject.toml
[project]
name = "dual-choice"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Pavel", email = "60060559+pyrogn@users.noreply.github.com" }
]
dependencies = [
    "pydantic>=2.7.0",
    "fastapi>=0.110.2",
    "uvicorn[standard]>=0.29.0",
    "slowapi>=0.1.9",
    "limits>=3.11.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.4",
]
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dual_choice"]
